xfsprogs: Add new name to attri/d
Source kernel commit: 4f0a04fd6dbddfbc483af57438c21ac50c26eb98

This patch adds two new fields to the atti/d.  They are nname and
nnamelen.  This will be used for parent pointer updates since a
rename operation may cause the parent pointer to update both the
name and value.  So we need to carry both the new name as well as
the target name in the attri/d.

This patch also applies the necassary updates to print the new
attri/d name fields.

Signed-off-by: Allison Henderson <allison.henderson@oracle.com>
Reviewed-by: Darrick J. Wong <djwong@kernel.org>